Tests: removed utils/ui/index.js file, re-organized imports
